faking iam and moving cdkApply* methods into their respective classes
