+++ lib/shared/policy-statement/9-final.ts
@@ -6,11 +6,11 @@ import { PolicyStatementWithPrincipal } from './6-principals';
 export class PolicyStatement extends PolicyStatementWithPrincipal {
   protected cdkApplied = false; // internally used to check if resources, actions and conditions have already been applied to the policy
 
+  //this method only exits to give it a final tsdoc, as it's exposed the the user.
   /**
    * JSON-ify the policy statement
    */
   public toJSON(): any {
-    //this method only exits to give it a final tsdoc
     return super.toJSON();
   }
 }
